sparkle: basic desired state RESTful API
The API now works, but does not perform any kind of access control.
Also, the entities are always returned in their fullest, which might
not always be what we want.

It also performs all the blocking DB operations in the main thread,
which won't scale at all.  This needs to be replaced with a thread
pool and some transaction ordering.

Signed-off-by: Jan Dvořák <mordae@anilinux.org>