place nic/bond/role rows even on evacuated hosts
We want to keep them connected to the network so that we can change our
mind later on and even if they are not available at the moment, there is
nothing to be changed about the placement.

Also, changes of the host itself change nothing about the placement of
these rows.

Signed-off-by: Jan Dvořák <mordae@anilinux.org>